third_party/WebKit/LayoutTests/platform/mac/external/wpt/IndexedDB/idbindex-rename-expected.txt

This is a testharness.js-based test.
FAIL IndexedDB index rename in new transaction assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL IndexedDB index rename in the transaction where it is created assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL IndexedDB index rename to the same name succeeds assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL IndexedDB index rename to the name of a deleted index succeeds assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL IndexedDB index swapping via renames succeeds assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL IndexedDB index rename stringifies non-string names assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL IndexedDB index can be renamed to "" assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL IndexedDB index can be renamed to "\u0000" assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL IndexedDB index can be renamed to "\uDC00\uD800" assert_equals: Expected success event, but got error event instead expected "success" but got "error"
Harness: the test ran to completion.
